feat: move lyrics to a different route
